Prep the post-ring caller ID on call set up as the dial command will change the called channel's caller ID again after initiating the dial. We can't rely on the caller ID after that.
The real problem is that a channel has only one caller ID associated
with it. When an incoming call triggers a dial on another channel the
caller ID of the incoming channel is copied to the called channel, which
then advertises it to the device it connects. But that causes a problem
if you do an attended transfer because when you put the incoming call on
hold and dial another channel _your_ channel has the wrong caller ID.
The "solution" that appears to have been used is to have dial set the
called channel's caller ID to the extension in the dialplan the dial
was called from. The observant will not that this is not the obviously
wrong caller ID from the incoming call but is still only vaguely right.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@3799 e56c2a57-502e-0410-a6e2-86b2f1ef08c6