RFC2833/4733 events must be clocked by the write side of the channel
Consider a call A -> B. The bridging watches A for incoming data and
writes each received packet (via translation, feature interpretation
etc.) to B. To fit event markers into the RTP stream leaving B they must
be clocked by the data being _written_ to B. Because the rtp stack is
not thread safe (and we have no wish to add locking on every frame
write) the events must be handled either by the same thread that is
writing audio data to B.

Hence RTP event markers must be generated and clocked by the writes of
audio data.

If there is no bridged source A or if A is using silence suppression
there will be no frames being written. Therefore we use the inband DTMF
tone generator to clock out silence frames for the period DTMF is to be
sent.

This provides an accurately timed stream of interleaved audio and event
packets. No strange jitter or discontinuities and overlaps between DTMF
events are not normally possible but are logged if they occur.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@4437 e56c2a57-502e-0410-a6e2-86b2f1ef08c6