Removed obsolete opbx_applystream.
The stream is bound to a channel when it's opened. We go to a lot of
trouble to choose the right instance of the file for the given stream.
If you decide you want it for a different stream after the fact you
should reopen it, not punt and pray.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@4441 e56c2a57-502e-0410-a6e2-86b2f1ef08c6