Avoid unnecessary mallocs, inline some simple and/or common stuff, remove some junk.
The trace-frames stuff is _gone_. It seemed to be some sort of
half-assed attempt to do some sort of leak detection but the list
pointers it used were the same ones used for frame queueing - and these
pointers get initialized to null and frame queueing code assumes they
are null to start with. In short, it didn't tell you anything useful (or
even correct) and there's a good chance enabling the trace-frames build
option would have caused more problems. There are better ways of finding
leaks!


git-svn-id: https://svn.callweaver.org/callweaver/trunk@4648 e56c2a57-502e-0410-a6e2-86b2f1ef08c6