Avoid using frames that have "disappeared".
A masquerade can happen in cw_write. A masquerade tears down the
translator and creates a new one (if necessary) on the modified channel.
cw_write is not a consumer of frames (it doesn't take a malloc'd frame
and free it when it's done) nor are translators generaters of malloc'd
frames (they pass references to private/stack data).

If a masquerade is triggered in cw_write by a frame from a translator
the frame is no longer valid after the masquerade because it exists in a
translator that has been destroyed.

Hence, cw_write must isolate is frame before performing a pending
masquerade. In turn this means that cw_write must take an extra level of
indirection on its frame argument and callers must be prepared for the
post-call frame reference to be different from the pre-call frame
reference they had before.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@4693 e56c2a57-502e-0410-a6e2-86b2f1ef08c6