Skip a leading "Digest" when checking an authorization response.
The previous check_auth changes to make tolerant of variable spacing
meant that the "Digest " lead in got prepended to the first item
rendering it unknown. In my environment the first item is "username" and
that is optional and just triggers a check against what we expected if
it is even present. So losing it didn't matter. Aastra phones however
send the response item first and losing that is, ah, "not helpful".

Before the variable spacing fixes it probably only worked because that
code split on _both_ commas _and_ spaces so the "Digest" lead in stood
(unintentionally) alone as an unknown, valueless item.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@4775 e56c2a57-502e-0410-a6e2-86b2f1ef08c6