Close ticket #402.
There was a race condition in NConference. When a member hangs up, the channel
structure may be deleted before the conference has time to remove it. Since it
uses the member->chan pointer to access some fields it may use random data.

So we now store a private copy of these data and use them when removing a
member. This will avoid the use of (possibly) stale data.



git-svn-id: https://svn.callweaver.org/callweaver/trunk@4974 e56c2a57-502e-0410-a6e2-86b2f1ef08c6