Make manager session descriptors non-blocking while handling requests.
We need the timeouts in cw_cli/cw_carefulwrite to work because there are
many manager actions and cli commands that write output while holding
locks. Even though request handling is in a separate thread this would
allow an external entity to stall parts of the core indefinitely.
Even with the timeout they can stall for up to 100ms (and repeatedly
for <100ms at a time).
This will (have to) change. But it will take a while...


git-svn-id: https://svn.callweaver.org/callweaver/trunk@5148 e56c2a57-502e-0410-a6e2-86b2f1ef08c6