CLI completion fixes and clean up
Readline is not broken - if you don't want it to think '@' breaks words
you have to tell it to use a non-default list of separators.
Completion generators no longer need to be called multiple times and 
malloc their returns, they can just write direct to the requesting
manager session's descriptor.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@5149 e56c2a57-502e-0410-a6e2-86b2f1ef08c6