Allow dialed peer to continue on the dialplan.
The default behaviour is to hangup the called peer when the caller hangs up.
This patch adds a new option to the Dial application
(J(context^extension^priority)), that will cause the dialed peer to jump to the
specified dialplan location and continue execution after the original caller
hangs up. This will only happen if the caller hangs up first.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@5288 e56c2a57-502e-0410-a6e2-86b2f1ef08c6