Use dynamic batching when writing CDRs
CDRs are queued as they become ready and written to the back end data sinks
by a thread. If the call rate exceeds the ability of the back ends to keep
up the CDR queue backlogs and the batch size passed to the back ends increases
thus allowing them to commit more CDRs per transaction and increase their
efficiency.

Dynamic batching writes CDRs as soon as possible while allowing scaling to
massive throughput.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5457 e56c2a57-502e-0410-a6e2-86b2f1ef08c6