Merge 1.2 r5181 (New function SIP_BUILD_DIAL(regex))
N.B. This is a bit sucky. There should be ways and means of iterating
over registries in generic ways in dialplan. Maybe that will wait
until we have interfaces to real languages though. Even so, you are
advised to expect to have to change things using SIP_BUILD_DIAL()
at some point in the future.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5496 e56c2a57-502e-0410-a6e2-86b2f1ef08c6