Copy any timestamp from the request to a "100 trying" response
RFC3261 8.2.6:
When a 100 (Trying) response is generated, any Timestamp header field
present in the request MUST be copied into this 100 (Trying) response.
If there is a delay in generating the response, the UAS SHOULD add a
delay value into the Timestamp value in the response.  This value MUST
contain the difference between the time of sending of the response and
receipt of the request, measured in seconds.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5598 e56c2a57-502e-0410-a6e2-86b2f1ef08c6