Clean up some non-essentials
Log a message if we abort a session because we were unable to write the
response to a request (and add a comment in case anyone thinks we could
just count the overflow and carry on).

No need to flip session fds in and out of blocking mode now output is
collected and encapsulated into a message rather than being spewed
directly at the fd.

Make sure there's a newline before we add the end-of-command-output
marker.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5609 e56c2a57-502e-0410-a6e2-86b2f1ef08c6