Clean up SIP retransmits
Remove entries from the retransmit queue as soon as we stop retransmits.
We don't need to wait for a final rather than a provisional response.

When discarding the retransmit queue in sip_destroy() we need to be sure
to actually free any queued messages as well as any auxiliary data they
own.


git-svn-id: https://svn.callweaver.org/callweaver/trunk@5642 e56c2a57-502e-0410-a6e2-86b2f1ef08c6