Use strcasecmp instead of strncasecmp when comparing arg strings in func_callerid.c
Someone seems to be confused by the use of 'n' functions that use
char arrays. The 'n' is NOT a magic way of removing security issues.
It has MEANING. In this case strcasecmp requires the strings to
match completely, strncasecmp requires only the first n characters to
match. strcasecmp is what is required here.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5647 e56c2a57-502e-0410-a6e2-86b2f1ef08c6