New syntax for Dial() options
Options may be expressed as individual arguments or may be grouped in a
single argument, the only rule being that, if you are using the (recommended)
Dial({...}, ...) syntax, an option that takes a (...) MUST always end a group.

That is to say: "abc", "a, b, c", "a, b(x), c", "ab(x), c" are all legal but
"ab(x)c" and "a, b(x)c" are only legal when you use the Dial(...&..., ...) syntax.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5664 e56c2a57-502e-0410-a6e2-86b2f1ef08c6