Make dialplan quoting work correctly
Previously it was possible for quotes embedded in a variable's value to
match with quotes enclosing the expansion, thus allowing a variable to
escape quoting and introduce its own comma separated terms that then get
treated as arguments in an outer function call.

Note that expansions that are NOT enclosed in quotes may still insert
args into an outer function call. This means that with:

	Set(myargs="arg1, arg2, arg3")

this:

	Func("${myargs}")

expands to a function call with a single argument consisting of the
string "arg1, arg2, arg3",

but this:

	Func(${myargs})

expands to a function call with three arguments.

It is likely that there will be a later patch to prevent expansions
producing more than a single argument whether quoted or not. Do not rely
on this behaviour. It exists solely for historical reasons.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5671 e56c2a57-502e-0410-a6e2-86b2f1ef08c6