$[ ... ]: Make <cond> ? <expr1> :: <expr2> only execute one of the exprs
Historically this has been broken in that it always executed both exprs
regardless. It only becomes a problem now we execute "bare" functions
(i.e. <func>(<args>) with no ${...} around it) while evaluating
expressions as the side effects of these can't be discarded so we must
be careful to only execute the expression required.

N.B. ${...} is expanded by variable substitution BEFORE $[...]
evaluation takes place. A direct func(args,...) WITHOUT ${...} round it
is expanded by the expression evaluation AFTER arg splitting. The
difference is subtle but very, very important to understand. Consider
the difference between:

    $[ ${cond} ? ${Goto(a)} :: ${Goto(b)} ]

and

    $[ ${cond} ? Goto(a) :: Goto(b) ]

Hint, one expands and executes both gotos before expression evaluation
is even attempted (and then uses the results of the goto calls as the
conditional values). The other tests the condition and executes just one
of the gotos.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5696 e56c2a57-502e-0410-a6e2-86b2f1ef08c6