pbx_ael should not abuse cw_expr() as a syntax checker
Expression evaluation is a two stage process, substitution then
evaluation. Just calling cw_expr() is simply wrong because it isn't
parsing what it would under true evaluation. Worse, you don't want to
expand or evaluate things with possible side effects when simply loading
a dialplan!

Removing this means that some syntax errors will only be detectable at
run-time rather than (possibly) load-time. But that's the nature of the
beast. Dialplan isn't a language. AEL is a text transformation tool that
transforms text into dialplan. AEL is not a language either. IMHO no one
should use AEL and it shouldn't even exist. At the very least it should
exist as a separate "compiler" and not be embedded.

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5714 e56c2a57-502e-0410-a6e2-86b2f1ef08c6