Add generic IP blacklisting functionality and plug in to chan_sip
The strategy is that requests that would create a new dialogue that come
from blacklisted IPs are silently ignored. If we do try the request then
if it is rejected (for reasons other than requiring an auth challenge)
we add the source IP to the blacklist with an initial duration of 60s.
Once the duration expires we remember the entry (but without blocking
requests) for a further 24 hours. If we re-add an IP for a remembered
entry we double the duration each time (subject to an optional maximum).

CLI commands are provided to add and remove entries, show entries and to
set the duration, maximum duration and remember times used by dynamic
blacklisting.