Reply to the address that a request was sent to
Previously we were going to the trouble of determining which of our many
aliases a message was sent to then ignoring it and going through a load
more work to figure which address to send replies from (possibly
incorrectly). Now we remember where the request was addressed to and use
that as the local address if a new dialogue needs to be created.