Add a SIPBLACKLIST function.
Use it as an invlaid handler to cause the IP trying to reach an invalid
number to be blacklisted in the same way bogus SIP registrations and
transactions are.

Really this should be a generic function but then we would need some way
to request the peer address (if any) from the channel's tech backend.
That will have to come later.