Tidy up retransmit code
We don't need to be able to claim sole use of a retrans message to be
able to set RTT. All we need is the reply and possibly the transmit time
if there is no timestamp header.

Remove unnecessary calls to retrans_stop_all()

Remove the need for retrans_stop*() to be given a dialogue as an
argument in preparation for moving to a more proper transaction layer.