Only discard *responses* with more than one Via
Requests can have more than one via. Proxies add vias as the request
goes out and strip them as the response comes back. The UAC sends a
request with one via and receives a response with one via. The UAS
receives a request with multiple vias and sends a response with multiple
vias.