Remove setoption method from channel drivers
Use control frames instead. In the process this removes the soft gain
adjustments. These were used inconsistently and were mainly there for
meetme's use. The implementation was certainly wrong. It can never be
right for an app or user to adjust the HW gain settings because that
could disrupt tone detection. Any adjustment for soft/loud voice has to
happen _after_ the DSP has done its work (or possibly as part of the DSP
processing?)