No need to schedule SIP dialogue destruction in many cases
The SIP transaction layer and object handling ensure that the dialogue
release happens at the appropriate time. We are therefore free to
initiate destruction per the UA flow and not worry about underlying
transaction requirements.

In the process we move the random data for auth nonces from the dialogue
to the peers and users. This means we do not have to create dialogues
before they are needed and prepares the way for handling out-of-dialogue
requests.